ti: vary tag color based on md5 hashing
each tag is assigned a color based on the last 6 digits in the md5 sum
of the tag. the color is blended with a base color weighting the base
color with a coefficient of 2.0 to have colors oscilliate around the
base.

each tag will therefore always have the same color, since the last
digits are used it is unlikely that similar tags (and adjacent in the
listing) will have the same color.